sends timestamped record to DB
